مترجم: إغلاق تباعُدات عدديّة قياسيّة موروثة (>> السالب، / و // عند INT64_MIN/-1) (#200)
المفسّر هو المرجع الدلاليّ في كلّ حالة:
- `>>` على السالب: كان يُرمَز CreateLShr (منطقيّ) ⇒ صار CreateAShr (حسابيّ يحفظ الإشارة) في arith_cmp.cpp + طيّ موقَّع int64_t >> في constant_folding_pass.cpp وsir_frontend_optimizer.cpp. (SHR يُنتَج حصرًا من >>؛ opcode SAR منفصل لتقليل القوّة — تحقّق Amelia.)
- `/` عند INT64_MIN/-1: كان safeRI=1 يُرجع INT64_MIN صحيحًا بينما المفسّر يرقّي إلى عشريّ؛ ضمّ minOverflow إلى isFloatRes في DIV (sad_dyn_repr.cpp).
- `//` عند INT64_MIN/-1: كان يُصلَّب Integer ⇒ SDiv فيض #DE ⇒ انهيار؛ إعادة توسيم نتيجة // على صحيحين إلى Any (expression_binary_op.cpp، نظير /) + minOverflow في FLOOR_DIV.
- `3==3.0`: كان مطابقًا أصلًا — اختبار تغطية فقط.

3 اختبارات تكافؤ مزدوج جديدة (SCALAR_shr_arithmetic، SCALAR_div_floordiv_int64min_overflow، SCALAR_mixed_int_float_compare).

البوّابات (بثنائيّات build/bin/Release الصريحة): P1 المزدوجة 171/171 (100%، +3 اختبارات) · gen --check أخضر. مراجعة عدائيّة (Amelia) 🟢 (~35 نقطة): AShr آمن (SHR من >> حصرًا)، إعادة توسيم // بلا انحدار صحّة (أثر أدائيّ فقط)، minOverflow دقيق (INT64_MIN و-1 حصرًا).

دَين موثَّق (خارج النطاق): `مطلق` للأعداد خارج 32-بت مؤجَّل — المفسّر نفسه معيب (toInt() 32-بت) فيتعذّر التكافؤ المزدوج دون إصلاح المفسّر (RFC عابر للمحرّكين).

+++ compiler/src/frontend/sir_frontend_optimizer.cpp
@@ -408,8 +408,15 @@ namespace Sad
                     constants[resultName] = SIROperand::ConstantI64(lhs.intValue << rhs.intValue);
                     return true;
                 case SIROpcode::SHR:
+                    // (AR) إزاحة يمنى حسابيّة (int64_t موقَّع) — عامل `>>` في المفسّر (المرجع)
+                    //      إزاحةٌ حسابيّة تحفظ الإشارة (-8 >> 1 = -4)، والخلفيّة تُصدر AShr.
+                    //      كان cast لـuint64_t يطوي إزاحةً منطقيّة فيباعد الطيَّ عن زمن التشغيل.
+                    // (EN) Signed (int64_t) arithmetic right shift — the interpreter (reference)
+                    //      `>>` is a sign-preserving arithmetic shift (-8 >> 1 = -4) and the
+                    //      backend emits AShr. The uint64_t cast folded a logical shift, diverging
+                    //      the fold from runtime.
                     constants[resultName] = SIROperand::ConstantI64(
-                        static_cast<int64_t>(static_cast<uint64_t>(lhs.intValue) >> rhs.intValue));
+                        lhs.intValue >> rhs.intValue);
                     return true;
 
                 default:
